Resolve style errors
Numerous style errors resolved with disregard for whether the code should be
rewritten or not.  One change was made which removes support for Python <2.6 in
order to fix the last issue.

As I think back to the work performed on this file, the thought of opening the
next one sends shivers down my spine.